Refactored cell magic to its own file; Implemented %%script and added test.
