+++ internal/nbtests/nbtests_test.go
@@ -661,58 +661,3 @@ func TestGonbui(t *testing.T) {
 	require.NoError(t, os.Remove(f.Name()))
 	clearNotebook(t, notebook)
 }
-
-// TestWritefile tests that `%%writefile` tests that `%%writefile` works
-func TestWritefile(t *testing.T) {
-	if testing.Short() {
-		t.Skip("Skipping integration (nbconvert) test for short tests.")
-		return
-	}
-	klog.Infof("GOCOVERDIR=%s", os.Getenv("GOCOVERDIR"))
-
-	// Create directory where to write the file, and set TEST_DIR env variable.
-	testDir, err := os.MkdirTemp("", "gonb_nbtests_writefile_")
-	require.NoError(t, err)
-	require.NoError(t, os.Setenv("TEST_DIR", testDir+"/"))
-	klog.Infof("TEST_DIR=%s/", testDir)
-
-	// Run notebook test.
-	notebook := "writefile"
-	f := executeNotebook(t, notebook)
-	err = Check(f,
-		Sequence(
-			Match(
-				OutputLine(1),
-				Separator,
-				fmt.Sprintf(`Cell contents written to "%s/poetry.txt".`, testDir),
-				Separator,
-			),
-			Match(
-				OutputLine(2),
-				Separator,
-				fmt.Sprintf(`Cell contents appended to "%s/poetry.txt".`, testDir),
-				Separator,
-			),
-		), *flagPrintNotebook)
-
-	require.NoError(t, err)
-	require.NoError(t, f.Close())
-	require.NoError(t, os.Remove(f.Name()))
-	clearNotebook(t, notebook)
-
-	// Checks the file was written.
-	filePath := path.Join(testDir, "poetry.txt")
-	contents, err := os.ReadFile(filePath)
-	require.NoErrorf(t, err, "Failed to read file %q that should have been written by the notebook.", filePath)
-	want := `Um trem-de-ferro é uma coisa mecânica,
-mas atravessa a noite, a madrugada, o dia,
-atravessou minha vida,
-virou só sentimento.
-
-Adélia Prado
-`
-	require.Equalf(t, want, string(contents), "Contents written to %q don't match", filePath)
-
-	// Remove directory.
-	require.NoError(t, os.RemoveAll(testDir))
-}
